Improved controllers users.js, registration still missing, formulated initial schemas for orders

+++ controllers/users.js
@@ -7,9 +7,11 @@ const responseUtils = require('../utils/responseUtils');
  * @param {http.ServerResponse} response
  */
 const getAllUsers = async response => {
-  response.writeHead(200, { 'Content-Type': 'application/json' });
-  console.log(await User.find({}));
-  response.end(JSON.stringify(await User.find({})));
+  //response.writeHead(200, { 'Content-Type': 'application/json' });
+  //console.log(await User.find({}));
+  //response.end(JSON.stringify(await User.find({})));
+  const users = await User.find({});
+  return responseUtils.sendJson(response, users);
 };
 
 /**
@@ -20,22 +22,23 @@ const getAllUsers = async response => {
  * @param {Object} currentUser (mongoose document object)
  */
 const deleteUser = async (response, userId, currentUser) => {
+  if (currentUser === null) {
+    return responseUtils.basicAuthChallenge(response);
+  } 
   const user = await User.findById(userId);
   if (user !== null) {
     if (userId.localeCompare(currentUser._id) == 0) {
-      responseUtils.badRequest(response, "Updating own data is not allowed");
-      }
-    else {
-      response.writeHead(200, { 'Content-Type': 'application/json' });
-      const user = await User.findById(userId);
+      return responseUtils.badRequest(response, "Updating own data is not allowed");
+    } else if (currentUser.role !== 'admin') {
+      return responseUtils.forbidden(response);
+    } else {
       await User.deleteOne({ _id: userId });
-      response.end(JSON.stringify(user));
+      return responseUtils.sendJson(response, user);
       }  
-    } 
-  else {
-    responseUtils.notFound(response);
-    }
-  };
+  } else {
+    return responseUtils.notFound(response);
+  }
+};
 
 /**
  * Update user and send updated user as JSON
@@ -47,7 +50,32 @@ const deleteUser = async (response, userId, currentUser) => {
  */
 const updateUser = async (response, userId, currentUser, userData) => {
   // TODO: 10.1 Implement this
-  };
+  if (currentUser === null) {
+    return responseUtils.basicAuthChallenge(resposne);
+  }
+  if (currentUser.role !== 'admin') {
+    return responseUtils.forbidden(response);
+  }
+  const user = await User.findById(userId);
+  if (user !== null) {
+    if (user._id !== currentUser._id) {
+      if (userData.role === 'customer' || userData.role === 'admin') {
+        user.role = userData.role;
+        await user.save();
+        return responseUtils.sendJson(response, user);
+      } else {
+        //role not allowed
+        return responseUtils.badRequest(response);
+      }
+    } else {
+      //can't update self
+      return responseUtils.badRequest(response, "Updating own data is not allowed")
+    }
+  } else {
+    //no defined user found
+    return responseUtils.notFound(response);
+  }
+};
 
 /**
  * Send user data as JSON
@@ -58,19 +86,19 @@ const updateUser = async (response, userId, currentUser, userData) => {
  */
 const viewUser = async (response, userId, currentUser) => {
   // TODO: 10.1 Implement this
-  /*
-  JOTENKIN TÄLLEEN:
+  
+  if (currentUser === null) {
+    return responseUtils.basicAuthChallenge(response);
+  } 
+  if (currentUser.role !== 'admin') {
+    return responseUtils.forbidden(response);
+  }
   const user = await User.findById(userId);
-  if (user !== null) {
-    const user = await User.findById(userId);
-    response.writeHead(200, { 'Content-Type': 'application/json' })
-    response.write(JSON.stringify(user));
-    response.end();
-    }
-  else {
-    responseUtils.notFound(response);
-    }
-   */ 
+  if (user === null) {
+    return responseUtils.notFound(response);
+  }
+  return responseUtils.sendJson(response, user);
+    
 };
 
 /**
